Fix: version to use snapshot, and update imports. Also did a little refactor to App4 em Movie class to make it working, the old version wasnt persisting in the database
