+++ scylla/pom.xml
@@ -28,9 +28,9 @@
     <dependencies>
 
         <dependency>
-            <groupId>org.eclipse.jnosql.mapping</groupId>
-            <artifactId>jnosql-cassandra-extension</artifactId>
-            <version>${project.version}</version>
+            <groupId>org.eclipse.jnosql.databases</groupId>
+            <artifactId>jnosql-cassandra</artifactId>
+            <version>${jnosql.version}</version>
         </dependency>
     </dependencies>
