+++ scylla/src/main/java/org/jnosql/demo/se/App3.java
@@ -13,10 +13,10 @@ package org.jnosql.demo.se;
 
 
 import com.datastax.oss.driver.api.core.ConsistencyLevel;
-import org.eclipse.jnosql.mapping.cassandra.column.CassandraTemplate;
-
 import jakarta.enterprise.inject.se.SeContainer;
 import jakarta.enterprise.inject.se.SeContainerInitializer;
+import org.eclipse.jnosql.databases.cassandra.mapping.CassandraTemplate;
+
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
